pyproject.toml
[project]
name = "quickspin-ai"
version = "0.1.0"
description = "Intelligent AI service for QuickSpin managed microservices platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "motor>=3.3.0",
    "httpx>=0.26.0",
    "langchain>=0.1.0",
    "langchain-groq>=0.0.1",
    "langgraph>=0.0.20",
    "chromadb>=0.4.22",
    "groq>=0.4.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "rich>=13.7.0",
    "kubernetes>=29.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[project.scripts]
quickspin-ai = "app.cli:app"
